+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ChangeLog
 =========
+1.5.1 (2020-03-31)
+-------------------------
+- Fix for listing clients
+- Fix for bookmark add response
+- Fix encoding site title
+- Fix get personal options
+
+
 1.5.0 (2020-03-24)
 -------------------------
 - Added support for MySQL